chore(terraform): add hostname configuration for Terraform remote state backend
The `main.tf` file was modified to add the `hostname` configuration for the Terraform remote state backend. The hostname is set to "spacelift.io".

The `prod.remote.tfbackend` and `staging.remote.tfbackend` files were also modified to add the `hostname` configuration for the respective Terraform workspaces. The hostname is set to "spacelift.io".