chore(workflows): update environment variables in promote-enterprisedocs-staging-to-production.yml, promote-staging-to-production.yml, send-docs-redirects-to-staging.yml, and send-enterprise-redirects-to-staging.yml
The environment variables `SPACELIFT_API_KEY_ENDPOINT`, `SPACELIFT_API_KEY_ID`, and `SPACELIFT_API_KEY_SECRET` have been added to the workflows `promote-enterprisedocs-staging-to-production.yml`, `promote-staging-to-production.yml`, `send-docs-redirects-to-staging.yml`, and `send-enterprise-redirects-to-staging.yml`. These variables are used to configure the Spacelift API key endpoint and credentials for the redirect creation scripts. This change allows for better management and security of the API key used in the workflows.