chore(workflows): update Terraform apply commands to disable locking for promoting staging to production and sending redirects to staging
The Terraform apply commands in the workflows `promote-enterprisedocs-staging-to-production.yml`, `promote-staging-to-production.yml`, `send-docs-redirects-to-staging.yml`, and `send-enterprise-redirects-to-staging.yml` have been updated to include the `-lock=false` flag. This change was made to disable locking during the apply process.

By disabling locking, it allows multiple Terraform runs to be executed concurrently without waiting for the lock to be released. This improves the efficiency and speed of the workflows, especially in scenarios where multiple deployments are happening simultaneously.

This change ensures that the Terraform apply commands can be executed without waiting for the lock to be released, reducing the overall deployment time and improving the workflow performance.