chore(workflows): add wait for concurrent workflows step to prevent race conditions
The workflows `promote-enterprisedocs-staging-to-production.yml`, `promote-staging-to-production.yml`, `send-docs-redirects-to-staging.yml`, and `send-enterprise-redirects-to-staging.yml` have been updated to include a new step that waits for concurrent workflows to finish before proceeding. This is done to prevent race conditions and ensure that the workflows are executed in the correct order. The `ahmadnassri/action-workflow-queue` action is used for this purpose.