chore(workflows): update AWS credentials configuration in promote-enterprisedocs-staging-to-production.yml, promote-staging-to-production.yml, send-docs-redirects-to-staging.yml, and send-enterprise-redirects-to-staging.yml
The AWS credentials configuration in the mentioned workflow files has been updated to use the `role-to-assume` parameter instead of `aws-access-key-id` and `aws-secret-access-key`. This change improves security by using AWS Identity and Access Management (IAM) roles for authentication instead of static access keys. The `role-to-assume` parameter specifies the IAM role to assume when configuring the AWS credentials.