DAT-11513      DevOps :: Enable OIDC for Github / AWS Integration (#218)
* chore(workflows): update AWS credentials configuration in promote-enterprisedocs-staging-to-production.yml, promote-staging-to-production.yml, send-docs-redirects-to-staging.yml, and send-enterprise-redirects-to-staging.yml

The AWS credentials configuration in the mentioned workflow files has been updated to use the `role-to-assume` parameter instead of `aws-access-key-id` and `aws-secret-access-key`. This change improves security by using AWS Identity and Access Management (IAM) roles for authentication instead of static access keys. The `role-to-assume` parameter specifies the IAM role to assume when configuring the AWS credentials.

* chore(redirect_creation): add new redirect for "/workflows/liquibase-community/using-the-run-job-pipeline-stage-with-spinnaker-test.html" to improve user experience and provide relevant information

* fix(workflows): update aws-actions/configure-aws-credentials version from v2 to v4 to ensure compatibility with the latest version

* chore(workflows): add write permissions for id-token and read permissions for contents in promote-enterprisedocs-staging-to-production.yml, promote-staging-to-production.yml, send-docs-redirects-to-staging.yml, and send-enterprise-redirects-to-staging.yml workflows

The permissions for the id-token and contents have been added to the workflows promote-enterprisedocs-staging-to-production.yml, promote-staging-to-production.yml, send-docs-redirects-to-staging.yml, and send-enterprise-redirects-to-staging.yml. This change is necessary to ensure that the workflows have the required permissions to perform write operations using the id-token and read operations using the contents.

* fix(workflows): add 'continue-on-error' flag to Terraform Show Plan step in send-docs-redirects-to-staging.yml and send-enterprise-redirects-to-staging.yml workflows to prevent workflow failure on error

* fix(workflows): update permissions for contents to write instead of read

fix(workflows): update permissions for contents to write instead of read in promote-staging-to-production.yml

fix(workflows): update permissions for contents to write instead of read in send-docs-redirects-to-staging.yml

fix(workflows): update permissions for contents to write instead of read in send-enterprise-redirects-to-staging.yml

fix(redirect_creation): remove unused redirect entry in variables.tf