FAQ grammar changes
Standardized the tenses in ran vs run.
Reduced the size of one of the sentences by removing unnecessary words.
Rewrote "Liquibase is better" for a couple of reasons;. 1. if you compare your product to others and state that it is better, legally you should have some data to back that up (independent surveys or reviews). 2. I was trying to sell Liquibase a little more than just saying "it's better". Pointing out that this feature is unique should prove to be a selling point.